reviews content broken up and complete
