+++ components/PageContent/ReviewsContent/index.ts
@@ -1,0 +1,3 @@
+import ReviewsContent from "./ReviewsContent";
+
+export default ReviewsContent;
